redux addingfavorites
In this exercise, you practiced setting up a new property as part of the state in the Redux store. You set up actions and a new reducer to update it, as well as updating the combineReducers function with the new reducer. Then you also updated the CampsiteInfo component to transfer the management of the favorite state from the local component state to the Redux store.